refactor mac settings
