+++ drivers/mac/menubar.go
@@ -11,46 +11,6 @@ import (
 	"github.com/pkg/errors"
 )
 
-// MenuBarConfig contains the menu bar configuration.
-type MenuBarConfig struct {
-	// The URL of the component to load in the menu bar.
-	// Set this to customize the whole menu bar.
-	//
-	// Default is mac.menubar.
-	URL string
-
-	// The URL of the app menu.
-	// Set this to customize only the app menu.
-	//
-	// Default is mac.appmenu.
-	AppURL string
-
-	// The URL of the edit menu.
-	// Set this to customize only the edit menu.
-	//
-	// Default is mac.editmenu.
-	EditURL string
-
-	// The URL of the window menu.
-	// Set this to customize only the window menu.
-	//
-	// Default is mac.windowmenu.
-	WindowURL string
-
-	// An array that contains additional menu URLs.
-	CustomURLs []string
-
-	// The URL of the help menu.
-	// Set this to customize only the help menu.
-	//
-	// Default is mac.helpmenu.
-	HelpURL string
-
-	// OnPreference is the function that is called when the Preference button
-	// from the default app menu is clicked.
-	OnPreference func()
-}
-
 func init() {
 	app.Import(&MenuBar{})
 	app.Import(&AppMenu{})
@@ -132,14 +92,12 @@ func (m *MenuBar) Render() string {
 
 // AppMenu is a component that describes the default app menu.
 type AppMenu struct {
-	AppName      string
-	OnPreference func()
+	AppName string
 }
 
 // OnMount initializes the menu application name.
 func (m *AppMenu) OnMount() {
 	m.AppName = app.Name()
-	m.OnPreference = driver.MenubarConfig.OnPreference
 	app.Render(m)
 }
 
@@ -149,18 +107,23 @@ func (m *AppMenu) Render() string {
 <menu>
 	<menuitem label="About {{.AppName}}" selector="orderFrontStandardAboutPanel:"></menuitem>
 	<menuitem separator></menuitem>
-	<menuitem label="Preferences…"
-			  keys="cmdorctrl+,"
-			  {{if .OnPreference}}onclick="OnPreference"{{end}}>
-	</menuitem>
+
+	<menuitem label="Preferences…" keys="cmdorctrl+," onclick="OnPreferences"></menuitem>
 	<menuitem separator></menuitem>
+
 	<menuitem label="Hide {{.AppName}}" keys="cmdorctrl+h" selector="hide:"></menuitem>
 	<menuitem label="Hide Others" keys="cmdorctrl+alt+h" selector="hideOtherApplications:"></menuitem>
 	<menuitem label="Show All" selector="unhideAllApplications:"></menuitem>
 	<menuitem separator></menuitem>
+
 	<menuitem label="Quit {{.AppName}}" keys="cmdorctrl+q" selector="terminate:"></menuitem>
 </menu>
 	`
+}
+
+// OnPreferences is the function called when the Preferences button is clicked.
+func (m *AppMenu) OnPreferences() {
+	app.Emit(PreferencesRequested, nil)
 }
 
 // EditMenu is a component that describes the default edit menu.
